Updated unit tests to use pylint infrastructure. Updated search procedure to use blosum max cutoff and to allow for vgene filtering by family or by family and by gene. Removed batch search procedure (for now).
