+++ src/antpack/database_tools/local_db_update_search.py
@@ -43,6 +43,7 @@ class LocalDBTool:
     def search(self, seq:str, annotation:tuple,
             mode:str="3", cdr_cutoff:float=0.25,
             blosum_cutoff:float=-1, max_cdr_length_shift:int=2,
+            use_vgene_family_only=True, symmetric_search=False,
             max_hits:int=1000, vgene:str="", species:str=""):
         """Searches the database and returns a list of nearest
         neighbors that meet the input criteria. The search can
@@ -70,7 +71,7 @@ class LocalDBTool:
                 and the maximum number of differences in cdrs 1 & 2
                 combined is 25% of their combined lengths rounded
                 to the nearest integer. cdr_cutoff is allowed to
-                range from 0 to 0.4 (if you want to use a larger
+                range from 0 to 0.33 (if you want to use a larger
                 value you will need to do a slower full-table scan
                 instead of using this function).
             blosum_cutoff (float): Either -1 or a value between 0 and
@@ -85,9 +86,23 @@ class LocalDBTool:
                 which the length of cdr3 for a match can differ from
                 the query. If 0, the results are required to have cdr3
                 be the same length as the query.
+            use_vgene_family_only (bool): If True, when filtering by
+                vgene, hits are required to have the same vgene *family*,
+                but can have a different vgene within that family. If
+                False, hits are required to have the same vgene AND
+                family. This argument is ignored if vgene or species is "".
+            symmetric_search (bool): If True, the cdr cutoff is applied
+                 both forward and reverse -- in other words, the Hamming
+                 distance from hit to query must be less than either
+                 cdr_cutoff * query_cdr_length OR cdr_cutoff * hit_cdr_length.
+                 This ensures that distance matrices constructed using search
+                 are symmetric, which is useful for clustering. If not planning
+                 to cluster or build a distance matrix, this is unnecessary.
             max_hits (int): The maximum number of hits to return.
             vgene (str): One of "" or a valid vgene. If a valid vgene,
-                hits are required to belong to the same v-gene family.
+                hits are required to match either the vgene family or
+                the specific vgene (depending on the use_vgene_family_only
+                argument).
             species (str): One of "" or a valid species (human, mouse,
                 alpaca, rabbit). If "", hits are required to belong to
                 the same assigned species.
@@ -102,84 +117,8 @@ class LocalDBTool:
         """
         return self.local_db_manager.search(seq, annotation,
                 mode, cdr_cutoff, blosum_cutoff,
-                max_cdr_length_shift, max_hits,
-                vgene, species)
-
-
-    def search_batch(self, seqs:list, annotations:list,
-            mode:str="3", cdr_cutoff:float=0.25,
-            blosum_cutoff:float=-1,
-            max_cdr_length_shift:int=2, max_hits:int=10,
-            vgenes:list=[], species:list=[]):
-        """Searches the database and returns a list of nearest
-        neighbors that meet the input criteria for each of the input
-        sequences. This is essentially a batch version of search.
-        The search can be conducted using the full sequence or a
-        sub-region of it (e.g. cdr3, the cdrs etc). This is faster
-        than doing individual searches if multithreading is in use
-        and you have a large number of queries.
-
-        Args:
-            seq (str): A list of input amino acid sequences. May contain
-                'X', must not contain '*'.
-            annotation (tuple): A list of tuples of (numbering, percent id,
-                chain type, error message) that is returned by
-                any of the analyze_seq functions for annotators
-                in AntPack.
-            mode (str): One of "3" or "123", indicating whether to
-                look for sequences that are similar based on cdr3 or
-                based on all three cdrs.
-            cdr_cutoff (float): The maximum number of differences
-                allowed from each query seq as a percentage of the
-                query seq cdr length(s) (excluding gaps). This
-                is assessed separately for cdr3 and cdrs 1/2 if
-                using mode "123". In other words, if searching
-                all three cdrs with cutoff 0.25, the maximum
-                number of differences allowed in cdr3 is 25%
-                of that cdr length rounded to the nearest integer,
-                and the maximum number of differences in cdrs 1 & 2
-                combined is 25% of their combined lengths rounded
-                to the nearest integer. cdr_cutoff is allowed to
-                range from 0 to 0.4 (if you want to use a larger
-                value you will need to do a slower full-table scan
-                instead of using this function).
-            blosum_cutoff (float): Either -1 or a value between 0 and
-                1 (inclusive). If < 0, this argument is ignored. If
-                between 0 and 1, any match that has BLOSUM mismatch score
-                less than (cdrlength * cutoff) for the query is
-                discarded. Specifying a value between 0 and 1 makes
-                the search more restrictive by excluding results that
-                are within CDR cutoff but have a highly improbable
-                mutation (as determined by BLOSUM scoring).
-            max_cdr_length_shift (int): The maximum +/- amount by
-                which the length of cdr3 for a match can differ from
-                a query. If 0, the results are required to have cdr3
-                be the same length as the query.
-            max_hits (int): The maximum number of hits to return for
-                each query.
-            vgene (str): Either an empty list or a list of valid vgenes
-                of the same length as seqs. If not empty, each hit is
-                required to belong to the same vgene family as the vgene
-                for that query.
-            species (str): Either an empty list or a list of valid species
-                of the same length as seqs. If not empty, each hit is
-                required to belong to the same species as the species
-                for that query.
-
-        Returns:
-            hits (list): A list of lists of sequence id numbers for hits
-                for each query. List[i] is the hits for query [i].
-                These can be used to retrieve the sequences and their
-                metadata using the get_sequence call.
-            distances (list): A list of lists of hamming distances of each
-                hit to the query, unless a blosum cutoff >= 0 was supplied,
-                in which case BLOSUM mismatch distance is calculated. List[i]
-                is the distance for query[i].
-        """
-        return self.local_db_manager.search_seqs(seqs, annotations,
-                mode, cdr_cutoff, blosum_cutoff,
-                max_cdr_length_shift, max_hits,
-                vgenes, species)
+                max_cdr_length_shift, use_vgene_family_only,
+                symmetric_search, max_hits, vgene, species)
 
 
     def get_sequence(self, seq_id:int):
