+++ test_requirements.txt
@@ -4,3 +4,4 @@ Biopython
 scipy
 blosum
 lmdb
+pytest
